[API] extract scope from application, domain, infrastructure
