+++ api/lib/scope-admin/domain/usecases/add-admin-role.js
